+++ lib/pdfserve/merge.rb
@@ -5,7 +5,7 @@ require "net/http"
 require "uri"
 require "json"
 
-module PdfserveClient
+module Pdfserve
   class Error < StandardError; end
 
   class Merge
