+++ lib/pdfserve/split.rb
@@ -5,37 +5,39 @@ require "net/http"
 require "uri"
 require "json"
 
-module PdfserveClient
+module Pdfserve
   class Error < StandardError; end
 
-  class Stamp
-    PATH = "/api/v1/pdf/stamp"
+  class Split
+    PATH = "/api/v1/pdf/split"
 
     def initialize(api_endpoint:, api_token: nil)
       @api_endpoint = api_endpoint + PATH
       @api_token = api_token
     end
 
-    def call(file_url, stamp_text)
+    def split(file_url:, pages:)
       uri = URI(api_endpoint)
+      uri.query = URI.encode_www_form({ pages: pages })
+
       http = Net::HTTP.new(uri.host, uri.port)
       http.use_ssl = uri.scheme == "https"
       request = Net::HTTP::Post.new(uri.request_uri)
 
       request["token"] = api_token unless api_token.nil?
-      form_data = [
-        ["files", file_url],
-        ["stamp_text", { "text" => stamp_text, "color" => "0,0,0", "position_name" => "tr", "over" => "true" }.to_json]
-      ]
+      form_data = [["splitfile", file_url]]
 
       request.set_form form_data, "multipart/form-data"
 
       response = http.request(request)
 
       if response.is_a?(Net::HTTPSuccess)
-        puts "Successful stamp!"
+        puts "Successful split!"
+        content_disposition = response.header["Content-Disposition"]
+        filename = content_disposition[/filename="?([^"]+)"?/, 1] || "archive.tar.gz"
+
         OpenStruct.new(
-          success: true, response: response.body, errors: ""
+          success: true, response: response.read_body, filename: filename, errors: ""
         )
       else
         puts "Failed!"
