+++ lib/pdfserve/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
-module PdfserveClient
+module Pdfserve
   VERSION = "2.0.0"
 end
